backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-agent-backend"
version = "0.0.0"
description = "Smart Agent Platform - Python Backend with FastAPI and LangGraph"
authors = [
    {name = "sumoer", email = "sumoqing2@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langgraph>=0.0.20",
    "mysql-connector-python>=8.2.0",
    "pymilvus>=2.3.4",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "httpx>=0.25.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"] 